Cleaned up all the warnings :)

+++ Source/SquidSalmple/EditManager/EditManager.h
@@ -11,7 +11,7 @@ public:
     void init (juce::ValueTree rootPropertiesVT);
 
     void loadBank (juce::File bankDirectory);
-    void loadChannel (juce::ValueTree squidChannelPropertiesVT, int channelIndex, juce::File sampleFile);
+    void loadChannel (juce::ValueTree squidChannelPropertiesVT, uint8_t channelIndex, juce::File sampleFile);
     void saveBank ();
 
 private:
